+++ abilian/services/auth/tests.py
@@ -3,6 +3,7 @@
 """
 from __future__ import absolute_import
 
+import json
 from abilian.testing import BaseTestCase
 from abilian.core.subjects import User
 
@@ -29,3 +30,21 @@ class TestAuth(BaseTestCase):
     rv = self.client.post('/user/login', data=kwargs)
     self.assertEquals(rv.status_code, 401, "expected 401, got:" + rv.status)
 
+
+  def test_api_post(self):
+    kwargs = dict(email=u'user@domain.tld', password='azerty', can_login=True)
+    u = User(**kwargs)
+    self.session.add(u)
+    self.session.commit()
+
+    rv = self.client.post('/user/api/login', data=json.dumps(kwargs),
+                          content_type='application/json')
+    self.assertEquals(rv.status_code, 200, "expected 200, got:" + rv.status)
+    self.assertEquals(rv.json, dict(email=u'user@domain.tld',
+                                    username=u'user@domain.tld',
+                                    fullname=u'Unknown',
+                                    next_url=u''))
+
+
+    rv = self.client.post('/user/api/logout')
+    self.assertEquals(rv.status_code, 200, "expected 200, got:" + rv.status)